requirements.txt
python==3.9.7
asgiref==3.4.1
better-profanity==0.7.0
certifi==2020.6.20
charset-normalizer==2.0.10
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
lxml==4.7.1
mysqlclient==2.1.0
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
wincertstore==0.2
